adjusted the code based on the comments
